don't create console pods until the balancer service is healthy
otherwise the console pods will crashloop for a bit on startup
