+++ src/orchestratord/src/controller/materialize/environmentd.rs
@@ -12,16 +12,13 @@ use std::{collections::BTreeMap, time::Duration};
 use anyhow::bail;
 use k8s_openapi::{
     api::{
-        apps::v1::{
-            Deployment, DeploymentSpec, DeploymentStrategy, RollingUpdateDeployment, StatefulSet,
-            StatefulSetSpec, StatefulSetUpdateStrategy,
-        },
+        apps::v1::{StatefulSet, StatefulSetSpec, StatefulSetUpdateStrategy},
         core::v1::{
             Capabilities, Container, ContainerPort, EnvVar, EnvVarSource, EphemeralVolumeSource,
-            HTTPGetAction, PersistentVolumeClaimSpec, PersistentVolumeClaimTemplate, Pod,
-            PodSecurityContext, PodSpec, PodTemplateSpec, Probe, SeccompProfile, SecretKeySelector,
-            SecretVolumeSource, SecurityContext, Service, ServiceAccount, ServicePort, ServiceSpec,
-            TCPSocketAction, Toleration, Volume, VolumeMount, VolumeResourceRequirements,
+            PersistentVolumeClaimSpec, PersistentVolumeClaimTemplate, Pod, PodSecurityContext,
+            PodSpec, PodTemplateSpec, Probe, SeccompProfile, SecretKeySelector, SecretVolumeSource,
+            SecurityContext, Service, ServiceAccount, ServicePort, ServiceSpec, TCPSocketAction,
+            Toleration, Volume, VolumeMount, VolumeResourceRequirements,
         },
         networking::v1::{
             IPBlock, NetworkPolicy, NetworkPolicyEgressRule, NetworkPolicyIngressRule,
@@ -60,9 +57,6 @@ pub struct Resources {
     persist_pubsub_service: Box<Service>,
     environmentd_certificate: Box<Option<Certificate>>,
     environmentd_statefulset: Box<StatefulSet>,
-    balancerd_external_certificate: Box<Option<Certificate>>,
-    balancerd_deployment: Option<Box<Deployment>>,
-    balancerd_service: Option<Box<Service>>,
 }
 
 impl Resources {
@@ -87,14 +81,6 @@ impl Resources {
         let environmentd_statefulset = Box::new(create_environmentd_statefulset_object(
             config, tracing, mz, generation,
         ));
-        let balancerd_external_certificate =
-            Box::new(create_balancerd_external_certificate(config, mz));
-        let balancerd_deployment = config
-            .create_balancers
-            .then(|| Box::new(create_balancerd_deployment_object(config, mz)));
-        let balancerd_service = config
-            .create_balancers
-            .then(|| Box::new(create_balancerd_service_object(config, mz)));
 
         Self {
             generation,
@@ -107,9 +93,6 @@ impl Resources {
             persist_pubsub_service,
             environmentd_certificate,
             environmentd_statefulset,
-            balancerd_external_certificate,
-            balancerd_deployment,
-            balancerd_service,
         }
     }
 
@@ -150,11 +133,6 @@ impl Resources {
 
         trace!("creating persist pubsub service");
         apply_resource(&service_api, &*self.persist_pubsub_service).await?;
-
-        if let Some(certificate) = &*self.balancerd_external_certificate {
-            trace!("creating new balancerd external certificate");
-            apply_resource(&certificate_api, certificate).await?;
-        }
 
         if let Some(certificate) = &*self.environmentd_certificate {
             trace!("creating new environmentd certificate");
@@ -292,20 +270,9 @@ impl Resources {
         namespace: &str,
     ) -> Result<(), anyhow::Error> {
         let service_api: Api<Service> = Api::namespaced(client.clone(), namespace);
-        let deployment_api: Api<Deployment> = Api::namespaced(client.clone(), namespace);
 
         trace!("applying environmentd public service");
         apply_resource(&service_api, &*self.public_service).await?;
-
-        if let Some(balancerd_deployment) = &self.balancerd_deployment {
-            trace!("creating new balancerd deployment");
-            apply_resource(&deployment_api, &*balancerd_deployment).await?;
-        }
-
-        if let Some(balancerd_service) = &self.balancerd_service {
-            trace!("creating new balancerd service");
-            apply_resource(&service_api, &*balancerd_service).await?;
-        }
 
         Ok(())
     }
@@ -1301,292 +1268,6 @@ fn create_environmentd_statefulset_object(
             ..mz.managed_resource_meta(mz.environmentd_statefulset_name(generation))
         },
         spec: Some(statefulset_spec),
-        status: None,
-    }
-}
-
-fn create_balancerd_external_certificate(
-    config: &super::Args,
-    mz: &Materialize,
-) -> Option<Certificate> {
-    create_certificate(
-        config.default_certificate_specs.balancerd_external.clone(),
-        mz,
-        mz.spec.balancerd_external_certificate_spec.clone(),
-        mz.balancerd_external_certificate_name(),
-        mz.balancerd_external_certificate_secret_name(),
-        None,
-    )
-}
-
-fn create_balancerd_deployment_object(config: &super::Args, mz: &Materialize) -> Deployment {
-    let security_context = if config.enable_security_context {
-        // Since we want to adhere to the most restrictive security context, all
-        // of these fields have to be set how they are.
-        // See https://kubernetes.io/docs/concepts/security/pod-security-standards/#restricted
-        Some(SecurityContext {
-            run_as_non_root: Some(true),
-            capabilities: Some(Capabilities {
-                drop: Some(vec!["ALL".to_string()]),
-                ..Default::default()
-            }),
-            seccomp_profile: Some(SeccompProfile {
-                type_: "RuntimeDefault".to_string(),
-                ..Default::default()
-            }),
-            allow_privilege_escalation: Some(false),
-            ..Default::default()
-        })
-    } else {
-        None
-    };
-
-    let pod_template_annotations = if config.enable_prometheus_scrape_annotations {
-        Some(btreemap! {
-            "prometheus.io/scrape".to_owned() => "true".to_string(),
-            "prometheus.io/port".to_owned() => config.balancerd_internal_http_port.to_string(),
-            "prometheus.io/path".to_owned() => "/metrics".to_string(),
-            "prometheus.io/scheme".to_owned() => "http".to_string(),
-        })
-    } else {
-        None
-    };
-    let mut pod_template_labels = mz.default_labels();
-    pod_template_labels.insert(
-        "materialize.cloud/name".to_owned(),
-        mz.balancerd_deployment_name(),
-    );
-    pod_template_labels.insert("app".to_owned(), "balancerd".to_string());
-    pod_template_labels.insert("materialize.cloud/app".to_owned(), mz.balancerd_app_name());
-
-    let ports = vec![
-        ContainerPort {
-            container_port: config.balancerd_sql_port,
-            name: Some("pgwire".into()),
-            protocol: Some("TCP".into()),
-            ..Default::default()
-        },
-        ContainerPort {
-            container_port: config.balancerd_http_port,
-            name: Some("http".into()),
-            protocol: Some("TCP".into()),
-            ..Default::default()
-        },
-        ContainerPort {
-            container_port: config.balancerd_internal_http_port,
-            name: Some("internal-http".into()),
-            protocol: Some("TCP".into()),
-            ..Default::default()
-        },
-    ];
-
-    let mut args = vec![
-        "service".to_string(),
-        format!("--pgwire-listen-addr=0.0.0.0:{}", config.balancerd_sql_port),
-        format!("--https-listen-addr=0.0.0.0:{}", config.balancerd_http_port),
-        format!(
-            "--internal-http-listen-addr=0.0.0.0:{}",
-            config.balancerd_internal_http_port
-        ),
-        format!(
-            "--https-resolver-template={}.{}.svc.cluster.local:{}",
-            mz.environmentd_service_name(),
-            mz.namespace(),
-            config.environmentd_http_port
-        ),
-        format!(
-            "--static-resolver-addr={}.{}.svc.cluster.local:{}",
-            mz.environmentd_service_name(),
-            mz.namespace(),
-            config.environmentd_sql_port
-        ),
-    ];
-
-    if issuer_ref_defined(
-        &config.default_certificate_specs.internal,
-        &mz.spec.internal_certificate_spec,
-    ) {
-        args.push("--internal-tls".to_owned())
-    }
-
-    let mut volumes = Vec::new();
-    let mut volume_mounts = Vec::new();
-    if issuer_ref_defined(
-        &config.default_certificate_specs.balancerd_external,
-        &mz.spec.balancerd_external_certificate_spec,
-    ) {
-        volumes.push(Volume {
-            name: "external-certificate".to_owned(),
-            secret: Some(SecretVolumeSource {
-                default_mode: Some(0o400),
-                secret_name: Some(mz.balancerd_external_certificate_secret_name()),
-                items: None,
-                optional: Some(false),
-            }),
-            ..Default::default()
-        });
-        volume_mounts.push(VolumeMount {
-            name: "external-certificate".to_owned(),
-            mount_path: "/etc/external_tls".to_owned(),
-            read_only: Some(true),
-            ..Default::default()
-        });
-        args.extend([
-            "--tls-mode=require".into(),
-            "--tls-cert=/etc/external_tls/tls.crt".into(),
-            "--tls-key=/etc/external_tls/tls.key".into(),
-        ]);
-    } else {
-        args.push("--tls-mode=disable".to_string());
-    }
-
-    let startup_probe = Probe {
-        http_get: Some(HTTPGetAction {
-            port: IntOrString::Int(config.balancerd_internal_http_port),
-            path: Some("/api/readyz".into()),
-            ..Default::default()
-        }),
-        failure_threshold: Some(20),
-        initial_delay_seconds: Some(3),
-        period_seconds: Some(3),
-        success_threshold: Some(1),
-        timeout_seconds: Some(1),
-        ..Default::default()
-    };
-    let readiness_probe = Probe {
-        http_get: Some(HTTPGetAction {
-            port: IntOrString::Int(config.balancerd_internal_http_port),
-            path: Some("/api/readyz".into()),
-            ..Default::default()
-        }),
-        failure_threshold: Some(3),
-        period_seconds: Some(10),
-        success_threshold: Some(1),
-        timeout_seconds: Some(1),
-        ..Default::default()
-    };
-    let liveness_probe = Probe {
-        http_get: Some(HTTPGetAction {
-            port: IntOrString::Int(config.balancerd_internal_http_port),
-            path: Some("/api/livez".into()),
-            ..Default::default()
-        }),
-        failure_threshold: Some(3),
-        initial_delay_seconds: Some(8),
-        period_seconds: Some(10),
-        success_threshold: Some(1),
-        timeout_seconds: Some(1),
-        ..Default::default()
-    };
-
-    let container = Container {
-        name: "balancerd".to_owned(),
-        image: Some(matching_image_from_environmentd_image_ref(
-            &mz.spec.environmentd_image_ref,
-            "balancerd",
-            None,
-        )),
-        image_pull_policy: Some(config.image_pull_policy.to_string()),
-        ports: Some(ports),
-        args: Some(args),
-        startup_probe: Some(startup_probe),
-        readiness_probe: Some(readiness_probe),
-        liveness_probe: Some(liveness_probe),
-        resources: mz.spec.balancerd_resource_requirements.clone(),
-        security_context: security_context.clone(),
-        volume_mounts: Some(volume_mounts),
-        ..Default::default()
-    };
-
-    let deployment_spec = DeploymentSpec {
-        replicas: Some(1),
-        selector: LabelSelector {
-            match_labels: Some(pod_template_labels.clone()),
-            ..Default::default()
-        },
-        strategy: Some(DeploymentStrategy {
-            rolling_update: Some(RollingUpdateDeployment {
-                // Allow a complete set of new pods at once, to minimize the
-                // chances of a new connection going to a pod that will be
-                // immediately drained
-                max_surge: Some(IntOrString::String("100%".into())),
-                ..Default::default()
-            }),
-            ..Default::default()
-        }),
-        template: PodTemplateSpec {
-            // not using managed_resource_meta because the pod should be
-            // owned by the deployment, not the materialize instance
-            metadata: Some(ObjectMeta {
-                annotations: pod_template_annotations,
-                labels: Some(pod_template_labels),
-                ..Default::default()
-            }),
-            spec: Some(PodSpec {
-                containers: vec![container],
-                node_selector: Some(
-                    config
-                        .balancerd_node_selector
-                        .iter()
-                        .map(|selector| (selector.key.clone(), selector.value.clone()))
-                        .collect(),
-                ),
-                security_context: Some(PodSecurityContext {
-                    fs_group: Some(999),
-                    run_as_user: Some(999),
-                    run_as_group: Some(999),
-                    ..Default::default()
-                }),
-                scheduler_name: config.scheduler_name.clone(),
-                service_account_name: Some(mz.service_account_name()),
-                volumes: Some(volumes),
-                ..Default::default()
-            }),
-        },
-        ..Default::default()
-    };
-
-    Deployment {
-        metadata: ObjectMeta {
-            ..mz.managed_resource_meta(mz.balancerd_deployment_name())
-        },
-        spec: Some(deployment_spec),
-        status: None,
-    }
-}
-
-fn create_balancerd_service_object(config: &super::Args, mz: &Materialize) -> Service {
-    let selector =
-        btreemap! {"materialize.cloud/name".to_string() => mz.balancerd_deployment_name()};
-
-    let ports = vec![
-        ServicePort {
-            name: Some("http".to_string()),
-            protocol: Some("TCP".to_string()),
-            port: config.balancerd_http_port,
-            target_port: Some(IntOrString::Int(config.balancerd_http_port)),
-            ..Default::default()
-        },
-        ServicePort {
-            name: Some("pgwire".to_string()),
-            protocol: Some("TCP".to_string()),
-            port: config.balancerd_sql_port,
-            target_port: Some(IntOrString::Int(config.balancerd_sql_port)),
-            ..Default::default()
-        },
-    ];
-
-    let spec = ServiceSpec {
-        type_: Some("ClusterIP".to_string()),
-        cluster_ip: Some("None".to_string()),
-        selector: Some(selector),
-        ports: Some(ports),
-        ..Default::default()
-    };
-
-    Service {
-        metadata: mz.managed_resource_meta(mz.balancerd_service_name()),
-        spec: Some(spec),
         status: None,
     }
 }
